### 2024-06-05 #6844 @editorialbot
[PRE REVIEW]: MocoExtendProblem: Interface Between OpenSim and MATLAB for Rapidly Developing Direct Collocation Goals in Moco
**Submitting author:** @Aravind-Sundararajan (<a href="http://orcid.org/0009-0001-6775-6596">Aravind Sundararajan</a>)
**Repository:** https://github.com/Aravind-Sundararajan/MocoExtendProblem
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
