+++ paper/paper.md
@@ -13,15 +13,15 @@ authors:
   - name: Aravind Sundararajan
     orcid: 0009-0001-6775-6596
     affiliation: 1
-	corresponding: true
+    corresponding: true
   - name: Varun Joshi
-	orcid: 0000-0003-3737-2302
+    orcid: 0000-0003-3737-2302
     affiliation: 2
   - name: Brian R. Umberger
-	orcid: 0000-0002-5780-2405
+    orcid: 0000-0002-5780-2405
     affiliation: 2
   - name: Matthew C. O'Neill
-	orcid: 0000-0001-9614-7813
+    orcid: 0000-0001-9614-7813
     affiliation: 1
 affiliations:
  - name: Department of Anatomy, Midwestern University, Glendale Arizona, USA
@@ -40,15 +40,15 @@ This repository features:
 
 - A `build.m` script that compiles goals in the `custom_goals` directory and procedurally constructs the C++/MATLAB class implementations and compiles the MEX interface.
 - Compatibility tested with OpenSim 4.2-4.5.
-	- OpenSim versions lower than 4.5 require unique modifications to the build pipeline since booleans for division by duration, distance and mass were migrated to the abstract MocoGoal. 
+  - OpenSim versions lower than 4.5 require unique modifications to the build pipeline since booleans for division by duration, distance and mass were migrated to the abstract MocoGoal. 
 - The ability to include `MEP` as a submodule, build, and use valid custom goals.
 - Three example custom goals in the `custom_goals` and `custom_goals_compat` directories.
 
 # Statement of need
 
-OpenSim is an open-source software platform for modeling musculoskeletal structures and creating dynamic simulations of movement [@Seth2018]. OpenSim enables researchers and clinicians to investigate how biological and non-biological structures respond to different loads, postures and activities in both static and dynamic situations. OpenSim has been used to study a wide range of biomechanical problems, such as the mechanics of walking and running [e.g. @Falisse2019], the impact of injury or disease on movement [e.g. @Johnson2022], and the effectiveness of rehabilitation exercises [e.g. @Spomer2023]. 
+OpenSim is an open-source software platform for modeling musculoskeletal structures and creating dynamic simulations of movement [@Seth2018]. OpenSim enables researchers and clinicians to investigate how biological and non-biological structures respond to different loads, postures and activities in both static and dynamic situations. OpenSim has been used to study a wide range of biomechanical problems, such as the mechanics of walking and running [e.g. @Falisse2019], the impact of injury or disease on movement [e.g. @Johnson2022], and the effectiveness of rehabilitation exercises [e.g. @Spomer2023].
  
-OpenSim Moco [@Dembia2020] employs an optimization paradigm called direct collocation to solve trajectory optimization problems that range from solving for muscle forces, to tracking experimental data, and fully predictive simulations. Direct collocation is a numerical optimal control method [@Kelly2017] that is computationally efficient and is used extensively in computational approaches to understanding biological movement. While direct collocation is powerful, Moco only provides a fixed set of optimization goals. It can be daunting for many users to develop custom goals in C++. We developed `MEP` so Moco users without experience compiling C++ can still write and test custom goals. 
[…]
