### 2024-06-05 fb7aa94
Update README.md
added badge for submission to JoSS status
