### 2024-06-06 #6844 @Kevin-Mattheus-Moerman
@Aravind-Sundararajan thanks for this submission. I will be the editor for this submission. I will soon start looking for reviewers. If you'd like to suggest reviewers that is fine/helpful. You can mention their GitHub handles here but leave out the @ symbol so they are not tagged.

For the moment, can  you:
- [ ] Check if any of those potentially missing DOIs can be addressed? It is okay if some truly do not have a DOI.
- [ ] Spell out USA as United States of America in your affiliations.
- [ ] Please rework/extend your README to clearly link to your license, the documentation, how to get started, how to test the package.  Here is an example for one of my MATLAB projects: https://github.com/gibbonCode/GIBBON
- [ ] Please add contributing guidelines to your project (examples here: https://contributing.md/), in addition a COC.md (code of conduct) file is recommended (see also: https://www.contributor-covenant.org/).
