### 2024-06-06 81e0822
adding Code of Conduct and Contributing pages
