### 2024-06-06 091d237
Update README.md to contain badge for GPL3.0 license
