### 2024-06-06 0c80347
WIP: improving README and including license/contributing/conduct files
