+++ README.md
@@ -1,8 +1,6 @@
 [![status](https://joss.theoj.org/papers/ab4ece70adece3811308955d52be6a2f/status.svg)](https://joss.theoj.org/papers/ab4ece70adece3811308955d52be6a2f)
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 
-# MocoExtendProblem
-
 ![skelemen](paper/skelemens.png)
 
 # Table of contents
@@ -15,17 +13,14 @@
 - [Code of conduct](#CodeOfConduct)  
 
 # Summary <a name="Summary"></a>
-
 MocoExtendProblem (`MEP`) is a framework to rapidly develop novel goals for biomechanical optimal control problems using OpenSim Moco and MATLAB (The Mathworks, Inc., Natick, MA, USA). `MEP` features several templates for testing and prototyping novel MocoGoals in lieu of rebuilding OpenSim or generating an .omoco file from C++ to load the problem into MATLAB. Instead, users structure custom goals, build them, and call custom goals from MATLAB scripts.
 
 # Setup <a name="Installation"></a>
-
 - Download and install OpenSim from [SimTK](https://simtk.org) and follow the documentation for setting up OpenSim’s MATLAB scripting environment.
 - Follow the instructions (OpenSim) to download necessary dependencies for both scripting in MATLAB and C++ development.
 - In MATLAB, configure MEX with mex -setup C++ to use the MS VisualStudio 2019+.
 
 # Getting started <a name="Start"></a>
-
 To create a new goal with `MEP`: 
 
 1. OpenSim 4.5+ users should copy a goal in the `custom_goals` directory while 4.2-4.4 users  should copy a goal in `custom_goals_compat`.
@@ -41,10 +36,17 @@ ep.addMocoCustomGoal('custom_goal',weight,power,divide_by_distance);
 ```
 
 # Testing <a name="Test"></a>
-blah blah blah
+In the test directory, we have provided some test scripts to be run with MATLAB desktop GUI:
+ - test_ExtendProblem_simple.m
+ - example2DWalking/WalkSim_Tracking.m
+ - example2DWalking/WalkSim_predictive.m
+ - driver.m
+
+Additionally if using an opensim version that is lower than 4.5, there are compatibility versions of `WalkSim_predictive.m` and `test_extendProblem_simple.m` to handle OpenSim version 4.2-4.4. The output of these scripts are compared against an OutputReference within the output directory. Note: you should stay on the top-level directory `MocoExtendProblem`
+
 # License <a name="License"></a>
 MEP is provided under:
-[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0). The [license file](https://github.com/gibbonCode/GIBBON/blob/master/LICENSE) is found on the GitHub repository.
+[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 
 # Contributing <a name="Contributing"></a>
 Please read our [contributing guidelines](CONTRIBUTING.md).
