+++ README.md
@@ -4,23 +4,27 @@
 ![skelemen](paper/skelemens.png)
 
 # Table of contents
-- [Project Summary](#Summary)  
-- [Setup](#Installation)  
-- [Getting started](#Start)
-- [Testing](#Test)
-- [License](#License)  
-- [Contributing](#Contributing)  
-- [Code of conduct](#CodeOfConduct)  
 
-# Summary <a name="Summary"></a>
+- [Summary](#summary)
+- [Setup](#setup)
+- [Getting Started](#getting-started)
+- [Testing](#testing)
+- [License](#license)
+- [Contributing](#contributing)
+- [Code of Conduct](#code-of-conduct)
+
+# Summary<a name="summary"></a>
+
 MocoExtendProblem (`MEP`) is a framework to rapidly develop novel goals for biomechanical optimal control problems using OpenSim Moco and MATLAB (The Mathworks, Inc., Natick, MA, USA). `MEP` features several templates for testing and prototyping novel MocoGoals in lieu of rebuilding OpenSim or generating an .omoco file from C++ to load the problem into MATLAB. Instead, users structure custom goals, build them, and call custom goals from MATLAB scripts.
 
-# Setup <a name="Installation"></a>
+# Setup<a name="setup"></a>
+
 - Download and install OpenSim from [SimTK](https://simtk.org) and follow the documentation for setting up OpenSim’s MATLAB scripting environment.
 - Follow the instructions (OpenSim) to download necessary dependencies for both scripting in MATLAB and C++ development.
 - In MATLAB, configure MEX with mex -setup C++ to use the MS VisualStudio 2019+.
 
-# Getting started <a name="Start"></a>
+# Getting Started<a name="getting-started"></a>
+
 To create a new goal with `MEP`: 
 
 1. OpenSim 4.5+ users should copy a goal in the `custom_goals` directory while 4.2-4.4 users  should copy a goal in `custom_goals_compat`.
@@ -35,21 +39,25 @@ ep = extend_problem(cptr);
 ep.addMocoCustomGoal('custom_goal',weight,power,divide_by_distance);
 ```
 
-# Testing <a name="Test"></a>
+# Testing<a name="testing"></a>
+
 In the test directory, we have provided some test scripts to be run with MATLAB desktop GUI:
- - test_ExtendProblem_simple.m
- - example2DWalking/WalkSim_Tracking.m
- - example2DWalking/WalkSim_predictive.m
- - driver.m
+- test_ExtendProblem_simple.m
+- example2DWalking/WalkSim_Tracking.m
+- example2DWalking/WalkSim_predictive.m
+- driver.m
 
 Additionally if using an opensim version that is lower than 4.5, there are compatibility versions of `WalkSim_predictive.m` and `test_extendProblem_simple.m` to handle OpenSim version 4.2-4.4. The output of these scripts are compared against an OutputReference within the output directory. Note: you should stay on the top-level directory `MocoExtendProblem`
 
-# License <a name="License"></a>
+# License<a name="license"></a>
+
 MEP is provided under:
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 
-# Contributing <a name="Contributing"></a>
+# Contributing<a name="contributing"></a>
+
 Please read our [contributing guidelines](CONTRIBUTING.md).
 
-# Code of conduct <a name="CodeOfConduct"></a>
+# Code of Conduct
+
 See [CODE_OF_CONDUCT](CODE_OF_CONDUCT.md)
