### 2024-06-06 5c5ac6d
Update CONTRIBUTING.md
removed todo sections in contributing.md for now
