### 2024-06-06 #6844 @Aravind-Sundararajan
Hello @Kevin-Mattheus-Moerman, thank you for taking the time to do this pre-review.

As for DOIs, [Mexplus](https://github.com/kyamagu/mexplus) is a project hosted on github, and the other two references are conference abstracts for presentations at podium or posters.

I have edited the authors list so the affiliations spell out United States of America.

I have also gone ahead and updated my readme and added the COC and CONTRIBUTING files. Let me know if there is anything else I can do here.

As far as reviewers, am I allowed to suggest people who did not yet volunteer? My thinking was to suggest users on [OpenSim](https://github.com/opensim-org/opensim-core)
