+++ utils/cpp_end.m
@@ -1,4 +1,11 @@
 function out = cpp_end()
+% cpp_end returns final lines for cpp class definition.
+%   out = cpp_end() 
+%
+%
+%   See also cpp_start, mex_start, mex_end, wrap_start, wrap_end
+
+
 out = "";
 %% CLASS DEF END
 class_def_end = [...
