+++ README.md
@@ -56,7 +56,7 @@ In the test directory, we have provided some test scripts to be run with MATLAB 
 - `example2DWalking/WalkSim_predictive.m`
 - `driver.m`
 
-Additionally if using an opensim version that is lower than 4.5, there are compatibility versions of `WalkSim_predictive.m` and `test_extendProblem_simple.m` to handle OpenSim version 4.2-4.4. The output of these scripts are compared against an OutputReference within the output directory. Note: you should stay on the top-level directory `MocoExtendProblem`
+Additionally if using an opensim version that is lower than 4.5, there are compatibility versions of `WalkSim_predictive.m` and `test_extendProblem_simple.m` to handle OpenSim version 4.2-4.4. The output of these scripts are compared against an OutputReference within the `MocoExtendProble\output` directory. Note: you should stay on the top-level directory `MocoExtendProblem`.
 
 # License<a name="license"></a>
 
