### 2024-06-12 #6844 @Kevin-Mattheus-Moerman
@chrisdembia @marcopus @Modenaxe @moorepants @jimmyDunne
would you be interested in reviewing this submission for [JOSS: The Journal of Open Source Software](https://joss.theoj.org), the title is: "_MocoExtendProblem: Interface Between OpenSim and MATLAB for Rapidly Developing Direct Collocation Goals in Moco_".

JOSS reviews happen here on GitHub and focus on [the software](https://github.com/Aravind-Sundararajan/MocoExtendProblem) as well as [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06844/joss.06844/10.21105.joss.06844.pdf). More on our review process can be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

If you are interested please let me know here, thanks!
