### 2024-08-14 #6844 @aasadi1
Hi Kevin,Sure, I can help with the review.Regards,AzarangOn Aug 12, 2024, at 05:35, Kevin Mattheus Moerman ***@***.***> wrote:﻿
@dgupta7 @aasadi1
would you be interested in reviewing this submission for JOSS: The Journal of Open Source Software, the title is: "MocoExtendProblem: Interface Between OpenSim and MATLAB for Rapidly Developing Direct Collocation Goals in Moco".
JOSS reviews happen here on GitHub and focus on the software as well as a short paper. More on our review process can be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
If you are interested please let me know here, thanks!
