### 2024-08-20 Aravind-Sundararajan/MocoExtendProblem#5 @0todd0000
Setup: mac unsupported?
The setup instructions on the [main repo page](https://github.com/Aravind-Sundararajan/MocoExtendProblem) do not indicate that Windows is required, but scripts appear to expect Windows. For example, the [build.m](https://github.com/Aravind-Sundararajan/MocoExtendProblem/blob/main/build.m) script yields warnings like this on `macos`:

```
Warning: Name is nonexistent or not a directory: /Users/USERNAME/MocoExtendProblem/\opensim 4.5\bin
```
which appears to mix macos and Windows-style paths

Is macos supported?  If not, no problem, please just state this in the Setup instructions and I'll switch over to Windows.
