### 2024-08-20 #7110 @0todd0000
@Kevin-Mattheus-Moerman Should I link to all issues in this thread or is it OK to just create issues and not reference them in this thread?  I have just submitted [a setup issue](https://github.com/Aravind-Sundararajan/MocoExtendProblem/issues/5) and it seems redundant to also mention this issue here. Please clarify, thanks!
