### 2024-08-20 Aravind-Sundararajan/MocoExtendProblem#5 @Aravind-Sundararajan
I had forgotten that SimTK allows you to download the pkg to install on macos. I did not develop with macos or linux distros in mind; however, msbuild through dotnet mono and cmake should be available to install, so technically, it should work if I change the paths, but I only have Matlab on Windows. Can you confirm that you have the required dependencies on macos:

- Matlab

- cmake

- msbuild

- any OpenSim version 4.2-4.5

When installed on mac, where are opensim headers installed? We can also ignore all this and say it only works on Windows.
