### 2024-08-27 Aravind-Sundararajan/MocoExtendProblem#5 @0todd0000
> Can you confirm that you have the required dependencies on macos

Yes, I do indeed have the required software.

> When installed on mac, where are opensim headers installed?

There are several header locations, but I think the main root in the default installation is:
```
/Applications/OpenSim 4.4/sdk
```
This is an alias for a folder inside the .app:
```
/Applications/OpenSim 4.4/OpenSim 4.4.app/Contents/Resources/opensim/sdk
```

> We can also ignore all this and say it only works on Windows.

I think this is a good idea. It may be best to limit support to Windows for purposes of review, as I don't think the supported platform(s) affect review outcome. So I'll shift over to Windows and retry setup. Please just indicate in the setup instructions that only Windows is explicitly supported at this stage.
