### 2024-08-27 7e5099b
Update README.md
chore: updating readme so it is clear that windows is required.
