### 2024-09-13 #7110 @dgupta7
Review of the written paper.
[Review_DG.pdf](https://github.com/user-attachments/files/16994230/Review_DG.pdf)
