### 2024-09-13 #7110 @dgupta7
Review of README.md
Note: I understand that some of the details I request here would be in the paper, but for a software, people will likely first read the ReadMe, so I assume that adding more details in ReadMe is important. Additionally, this comment is based only on reading the ReadMe file. I haven't started the installation process yet.

Summary:
- is “in lieu of” followed by “instead” grammatically accurate?

Setup & Requirements:
- Does MEP work for OpenSim versions below 4.2? please specify.
- Can you add which versions of visual studio and CMake was MEP tested on? Visual studio gives options to install multiple different packages. Which ones should be install?
- Please add a link to the “documentation for setting up OpenSim’s MATLAB scripting environment”
- Add link to the instructions that you are referring to in “Follow the instructions (OpenSim) to download necessary dependencies for both scripting in MATLAB and C++ development”.
- In MATLAB, configure MEX with mex -setup C++ to use the MS VisualStudio 2019+. Please specify clearly how to do this. Should the user just copy paste mex -setup C++ in command prompt?
- The figure does not belong in the Setup & Requirements section.

Getting Started:
Creating MEX interface:
- Is running the build.m a one-time step or do we need to do this every time we open MATLAB?
- “you are free to add the bin\relwithdebinfo to your matlab path.” Please be specific, do you want the users to do this? What is the advantage of doing this?
- I am confused. Should one fork the github repository? Or use the build.m and then add bin\relwithdebinfo to MATLAB path?
Creating a new goal:
- When you say “copy a goal”, what do you mean? Is it a .m file that is to be copied? I think it is a folder?
- “Replace mentions of the original goal name to that of your new custom goal name in each of the 5 files and file names, being careful to also modify the include guards in the dll and register types header files.“. Just by reading this, it is not clear. What do you mean by “being careful to also modify the include guards in the dll and register types header files”. How do I do that? .dll does not open in Notepad++, so do I need another software? Maybe you could also add pictures or links to pictures.
- I think that step 1 and 2 of Creating a new goal can be automated. It seems to me that the only thing happening here is just copying some things and changing names of files. I think this automation will greatly help users.
- By “Reimplement”, do you mean overwrite? But where do these functions (constructProperties(), initializeOnModelImpl(), calcIntegrandImpl(), calcGoalImpl()) live? Are they in one of the .m files? In which folder?
- I am not a C coding expert, so I do not understand “To incorporate extend_problem goals into an existing MATLAB script, a C-style pointer to the instantiated MocoProblem is passed as a constructor argument to the extend_problem.m class that wraps the MEP MEX. Class methods of extend_problem.m are then used to add custom goals to the MocoProblem.” I think it is okay to keep it, but you should also provide a layman explanation.
- I do not understand what I need to do with the 3 lines of code. Do I need to run them? When and where do I run them? Are they an example for creating a goal? It seems that it needs the variable problem, so I am guessing there are things that are run before these lines. What are those things.

Testing:
- What do each of these tests do? Do they create a new goal? What is the new goal that they create? Do they also run a simulation? Is it a walking simulation?
- Maybe the figure in the Setup & Requirements section belongs here, but it would need explanations.
- Where does the reference come from?
- I am confused with the “Note: you should stay on the top-level directory MocoExtendProblem.”. But the testing files are all in the test directory.

License:
I am wondering GPL license is being used and not MIT? MIT gives more freedom to the users.
