### 2024-09-13 #7110 @dgupta7
When I started installing the software, I typed mex -setup C++ in MATLAB command prompt and it gave me the error "Error using mex
Supported compiler not found. You can install the
freely available MinGW-w64 C/C++ compiler; see
Install MinGW-w64 Compiler. For more options,
visit https://www.mathworks.com/support/compilers."

I am wondering what is the reason. Maybe because while installing visual studio I only installed the "Desktop development with C++" workload. Do we need more than that?
![VS](https://github.com/user-attachments/assets/9591fc59-9c0b-410a-9c56-3b53ff406a46)
