### 2024-09-13 #7110 @Aravind-Sundararajan
> When I started installing the software, I typed mex -setup C++ in MATLAB command prompt and it gave me the error "Error using mex Supported compiler not found. You can install the freely available MinGW-w64 C/C++ compiler; see Install MinGW-w64 Compiler. For more options, visit https://www.mathworks.com/support/compilers."
>
> I am wondering what is the reason. Maybe because while installing visual studio I only installed the "Desktop development with C++" workload. Do we need more than that? ![VS](https://private-user-images.githubusercontent.com/85564166/367317533-9591fc59-9c0b-410a-9c56-3b53ff406a46.JPG?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjYyMzkwNDAsIm5iZiI6MTcyNjIzODc0MCwicGF0aCI6Ii84NTU2NDE2Ni8zNjczMTc1MzMtOTU5MWZjNTktOWMwYi00MTBhLTljNTYtM2I1M2ZmNDA2YTQ2LkpQRz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MTMlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTEzVDE0NDU0MFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWNjNTllZTUzYjc2NjVmNTI3M2VhMmZlYWJiMmRlOTQ4YmQ5OTc2MjdkZTYyN2MyMTM1MzliMmRmMTk1NjVjYzcmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.kf8ZLhfNaBfwtNOt6ue4_GPfjD07SYkK7oisMiHp5uo)

that should be enough. Ensure that cl.exe can be run from cmd or powershell, or that your vs version is added to the system PATH environment variable
