### 2024-09-13 0b164c0
chore: Update README.md instructions
