### 2024-09-15 c576539
WIP: updating readme for clarity
