+++ README.md
@@ -16,30 +16,29 @@
 
 # Summary<a name="summary"></a>
 
-MocoExtendProblem (`MEP`) is a framework to rapidly develop novel goals for biomechanical optimal control problems using OpenSim Moco and MATLAB (The Mathworks, Inc., Natick, MA, USA). `MEP` features several templates for testing and prototyping novel MocoGoals in lieu of rebuilding OpenSim or generating an .omoco file from C++ to load the problem into MATLAB. Instead, users structure custom goals, build them, and call custom goals from MATLAB scripts.
+MocoExtendProblem (`MEP`) is a framework to rapidly develop novel goals for biomechanical optimal control problems using OpenSim Moco and MATLAB (The Mathworks, Inc., Natick, MA, USA). `MEP` features several templates for testing and prototyping novel MocoGoals. In lieu of rebuilding OpenSim or generating an .omoco file from C++ to load the problem into MATLAB, users structure custom goals, build them, and call custom goals from MATLAB scripts.
 
 # Setup and Requirements<a name="setup"></a>
 - `MEP` was developed with Windows in mind, MacOS and linux are not supported at this time.
-- `MEP` runs on MATLAB (tested on 2022a and above) and requires [visual studio](https://visualstudio.microsoft.com) as well as [CMake](https://cmake.org/download/)
-- Download and install OpenSim from [SimTK](https://simtk.org) and follow the documentation for setting up OpenSim’s [MATLAB scripting environment](https://opensimconfluence.atlassian.net/wiki/spaces/OpenSim/pages/53089380/Scripting+with+Matlab).
+- `MEP`supports OpenSim versions 4.5 and has compatability for OpenSim 4.2-4.4.
+- `MEP` runs on MATLAB (tested on 2022a and above) and was tested with [visual studio 2019+](https://visualstudio.microsoft.com) as well as [CMake 3.23.3+](https://cmake.org/download/)
+- Download and install an OpenSim version from 4.2-4.5 from [SimTK](https://simtk.org) and follow the documentation for setting up OpenSim’s [MATLAB scripting environment](https://opensimconfluence.atlassian.net/wiki/spaces/OpenSim/pages/53089380/Scripting+with+Matlab).
 - Follow the instructions (OpenSim) to download necessary dependencies for both MATLAB Scripting and C++ development.
 - Be sure to add visual studio to your system PATH environment variable such that `cl.exe` can be run from command prompt or powershell.
-- In MATLAB, configure MEX with mex -setup C++ to use the MS VisualStudio 2019+.
-
-![Max Coordinate Goal](paper/MEP_point_mass_max.png)
+- In MATLAB, configure MEX by running `mex -setup C++` in the MATLAB command window to use MS VisualStudio 2019+.
 
 # Getting Started<a name="getting-started"></a>
 
 ## Compile MEX interface
 
-From the top-level directory (`MocoExtendProblem`) there is a `build.m` script. Running this script will generate the `ExtendProblem` class and the MEX interface. NOTE: building and testing the class requires being in the top-level directory; however after the build is successful, you are free to add the `bin\relwithdebinfo` to your matlab path.
+From the top-level directory (`MocoExtendProblem`) there is a `build.m` script. Running this script will regenerate the `ExtendProblem` class and the MEX interface. Each time any goal implementations need to be changed/updated, build.m should be run. NOTE: building and testing the class requires being in the top-level directory (`MocoExtendProblem`); however, after the build is successful, you are free to add the `bin\relwithdebinfo` to your matlab path so you can access `MEP` .
 
-Forking and adding `MEP` as a submodule to your project and then building is the preferred method of including it to a new or existing project.
[…]
