+++ README.md
@@ -64,7 +64,7 @@ Additionally, if using an opensim version that is lower than 4.5, there are comp
 # License<a name="license"></a>
 
 MEP is provided under:
-[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+[![License: MIT](https://img.shields.io/badge/license-MIT-blue.svg)](https://opensource.org/license/mit)
 
 # Contributing<a name="contributing"></a>
 
