### 2024-09-15 ac20e8e
chore: support msg in readme
