### 2024-09-16 #7110 @Aravind-Sundararajan
> How do I ensure that cl.exe can be run from cmd? or how do I ensure that vs version is added to the system PATH environment variable? My VS 2022 community does not have a clear bin folder to add to the system paths. Another thing I tried was to start matlab from x64 Native Command Prompt, and then run the build.m in MEP, but it still gave me the error that it cannot find the compiler.
>
> It is possible that I am doing something stupid but I think that installation problems will be common among users and more clear documentation here is needed to help users with exactly what to do step-by-step to get MEP setup correctly.

you can check in verbose mode where matlab is looking for vs2019 or 2022, `mex -setup CPP -v`. I don't remember if you need to install a package for mex support when installing matlab so it's possible something is missing that I don't know.

I have this in my path environment variable `C:\Program Files\Microsoft Visual Studio\2022\Community\MSBuild\Current\Bin`
