### 2024-09-18 #7110 @aasadi1
I have the same problem. VS is in my path environment: C:\Program Files\Microsoft Visual Studio\2022\Community\Msbuild\Current\Bin

but yet when I run cl.exe in the terminal I get this error:

cl.exe : The term 'cl.exe' is not recognized as the name of a cmdlet, function, script file, or operable program.
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
At line:1 char:1
+ cl.exe
+ ~~~~~~
    + CategoryInfo          : ObjectNotFound: (cl.exe:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException
