### 2024-09-18 #7110 @aasadi1
I downloaded CMake and installed it an rant build.m and now I get this error:

CMake Deprecation Warning at CMakeLists.txt:1 (cmake_minimum_required):
  Compatibility with CMake < 3.5 will be removed from a future version of
  CMake.

  Update the VERSION argument <min> value or use a ...<max> suffix to tell
  CMake that the project does not need compatibility with older versions.

CMake Error at CMakeLists.txt:2 (project):
  Running

   'nmake' '-?'

  failed with:

   no such file or directory

CMake Error: CMAKE_C_COMPILER not set, after EnableLanguage
CMake Error: CMAKE_CXX_COMPILER not set, after EnableLanguage
-- Configuring incomplete, errors occurred!
MSBuild version 17.11.9+a69bbaaf5 for .NET Framework
MSBUILD : error MSB1009: Project file does not exist.
Switch: C:\Users\azara\Downloads\MocoExtendProblem-main\MocoExtendProblem-main\build\customGoals.sln
detected using opensim 4.4 or lower.
Error using fprintf
Invalid file identifier. Use fopen to generate a valid file identifier.

Error in build_extend_class (line 29)
fprintf(fid,'%s\n',s);

Error in build (line 44)
build_extend_class(fullfile(bindir,config,cppName),fullfile(bindir,config,wrapName), opensim_install);

Also I changed OpenSim 4.5 to 4.4 in the code.
