### 2024-09-18 #7110 @Aravind-Sundararajan
gi

> I downloaded CMake and installed it an rant build.m and now I get this error:
>
> CMake Deprecation Warning at CMakeLists.txt:1 (cmake_minimum_required): Compatibility with CMake < 3.5 will be removed from a future version of CMake.
>
> Update the VERSION argument value or use a ... suffix to tell CMake that the project does not need compatibility with older versions.
>
> CMake Error at CMakeLists.txt:2 (project): Running
>
> 'nmake' '-?'
>
> failed with:
>
> no such file or directory
>
> CMake Error: CMAKE_C_COMPILER not set, after EnableLanguage CMake Error: CMAKE_CXX_COMPILER not set, after EnableLanguage -- Configuring incomplete, errors occurred! MSBuild version 17.11.9+a69bbaaf5 for .NET Framework MSBUILD : error MSB1009: Project file does not exist. Switch: C:\Users\azara\Downloads\MocoExtendProblem-main\MocoExtendProblem-main\build\customGoals.sln detected using opensim 4.4 or lower. Error using fprintf Invalid file identifier. Use fopen to generate a valid file identifier.
>
> Error in build_extend_class (line 29) fprintf(fid,'%s\n',s);
>
> Error in build (line 44) build_extend_class(fullfile(bindir,config,cppName),fullfile(bindir,config,wrapName), opensim_install);
>
> Also I changed OpenSim 4.5 to 4.4 in the code.
![image](https://github.com/user-attachments/assets/07918429-6d69-4c55-ad56-5a6606ec4167)

looks like it wants to use nmake makefiles instead of Visual Studio 2022 or 2019, if you call `cmake -G` you'll get a list of generators available.

you can try changing the call on line 29 to
```
system("cmake CmakeLists.txt -S . -B """+builddir+""" -DOSim_Version=4 -G 'Visual Studio 17 2022'");
```
thereby specifying the generator. There are also ways to change the default generator too but you'll have to look that up as I'm not too familiar.
