### 2024-09-18 #7110 @Aravind-Sundararajan
> just did:
>
> Unrecognized function or variable 'builddir'.
>
> Error in build_extend_class (line 31) system("cmake CmakeLists.txt -S . -B """+builddir+""" -DOSim_Version=4 -G 'Visual Studio 17 2022'");
>
> Error in build (line 44) build_extend_class(fullfile(bindir,config,cppName),fullfile(bindir,config,wrapName), opensim_install);

I think you've changed a line in the wrong file, you should stash changes and pull again as I have updated the repo with the desired change
