### 2024-09-18 #7110 @aasadi1
I finally got it installed. Maybe you wanna add this part to the readme file for people like me who are not familiar with software development.

Open Visual Studio
Go to Tools -> Get Tools and Features
In the "Workloads" tab enable "Desktop development with C++"
Click Modify at the bottom right

Also I think you should ask the user to change the VS and OpenSim version in the build.m file like we did.
