### 2024-09-19 #7110 @dgupta7
I was able to run build.m on a different computer. I actually did not need VS bin folder in system PATH and cl.exe still does not run from command prompt. Are those actually required for build.m, or maybe needed later?
One notable difference was that on this other computer I used MATLAB 2022b, I was using 2021a earlier. MEX -setup C++ itself was not working on the computer with 2021a. Does MATLAB version matter for MEP? if yes, please add that to the documentation.
