### 2024-09-23 f6f64a2
made instructions for vs workload more clear
