### 2024-09-23 bb266df
remove line about cl.exe from readme.md
