### 2024-09-23 83cb814
added a template.m script that helps make the 5 goal implementation files and the goaldir
