### 2024-09-30 043006e
WIP: rough updates for paper and bib
