+++ paper/paper.bib
@@ -13,7 +13,16 @@
     abstract = { This paper is an introductory tutorial for numerical trajectory optimization with a focus on direct collocation methods. These methods are relatively simple to understand and effectively solve a wide variety of trajectory optimization problems. Throughout the paper we illustrate each new set of concepts by working through a sequence of four example problems. We start by using trapezoidal collocation to solve a simple one-dimensional toy problem and work up to using Hermite--Simpson collocation to compute the optimal gait for a bipedal walking robot. Along the way, we cover basic debugging strategies and guidelines for posing well-behaved optimization problems. The paper concludes with a short overview of other methods for trajectory optimization. We also provide an electronic supplement that contains well-documented MATLAB code for all examples and methods presented. Our primary goal is to provide the reader with the resources necessary to understand and successfully implement their own direct collocation methods. (An erratum is attached.) }
 }
 
-
+@article{delp2007,
+author = {Delp, Scott and Anderson, Frank and Arnold, Allison and Loan, Peter and Habib, A. and John, Chand and Guendelman, Eran and Thelen, Darryl},
+year = {2007},
+month = {12},
+pages = {1940 - 1950},
+title = {OpenSim: Open-Source Software to Create and Analyze Dynamic Simulations of Movement},
+volume = {54},
+journal = {Biomedical Engineering, IEEE Transactions on},
+doi = {10.1109/TBME.2007.901024}
+}
 
 
 @article{Dembia2020,
