+++ README.md
@@ -61,6 +61,130 @@ ep = extend_problem(cptr);
 ep.addMocoCustomGoal('custom_goal',weight,power,divide_by_distance);
 ```
 
+## Tutorial
+
+
+first clone MocoExtend Problem.
+```
+git clone git@github.com:Aravind-Sundararajan/MocoExtendProblem.git
+```
+Next, open MATLAB and run build.m to see that your setup is working.
+![image](https://github.com/user-attachments/assets/7ec8c957-8abd-45d4-a5cc-fc76f8eb5965)
+
+After building, you should see the following in the command window:
+![image](https://github.com/user-attachments/assets/2bc359da-b3f5-45a7-881b-bf6b8a8180f0)
+
+To build a new goal, we are going to copy a goal folder to serve as a template, I'll call this new goal MocoCustomGoal.
+
+![image](https://github.com/user-attachments/assets/a2182699-0632-49c8-9b7a-1d395ff15239)
+
+![image](https://github.com/user-attachments/assets/96c63e1b-160b-4175-b5cb-1f299228c5f6)
+
+We need to change the naming conventions of the 5 files (GoalName.cpp, GoalName.h, osimGoalNameDLL.h, RegisterTypes_osimGoalName.cpp, RegisterTypes_osimGoalName.h) from MocoActivationGoal to MocoCustomGoal
+You can do this by hand or using an ide/editor like vscode or notepad++.
+
+![image](https://github.com/user-attachments/assets/c9b46c1e-4f7a-4004-8230-c40b68cc20a7)
+
+We also need to change the  naming conventions within each of the files. This is relatively easy to regex replace or find&replace with vscode, but also not challenging by hand either. I am going to do this with vscode. You should have 45 replacements across the 5 files. (Note I am checking the box for Preserve Case so I can change the class names and includeguards in one go.)
+
+![image](https://github.com/user-attachments/assets/9f75691b-d203-47b9-991a-a3816ef68d0f)
+
+Now let's run build.m again and make sure our new MocoCustomGoal works.
+
+![image](https://github.com/user-attachments/assets/7ef1dc71-4a3b-4b58-8b10-7ebf3e9b1d34)
+
+If we look in the bin/RelWithDebInfo directory, we will see that our new custom goal has been created.
+
+![image](https://github.com/user-attachments/assets/31809208-2d24-4a49-a17c-2246cfa2f397)
+
+And if we check `bin/RelWithDebInfo/extend_problem.m` you will see we have a new method for adding the MocoCustomGoal to a MocoProblem.
+
+![image](https://github.com/user-attachments/assets/61de7002-44a4-4d43-bdc6-6988eaa57b48)
+
+Let's try adding MocoCustomGoal to a problem. Quickly we can edit `test\test_ExtendProblem_simple.m`
+
+Comment line 81, and add a new line `ep.addMocoCustomGoal('custom_goal',1.0, true, false, false, false);` To know what the boolean arguments do, see the interface that we just lookled at (`bin/RelWithDebInfo/extend_problem.m`)
+
+![image](https://github.com/user-attachments/assets/f074e409-72a2-47e4-9be9-470ae1c82af0)
+
+Now if we hit run, and Add to Path, we should see 
+
+```
+Adding MocoCustomGoal goal
+
+List of user-set options:
+
+                                    Name   Value                used
+                   hessian_approximation = limited-memory        yes
+                      print_user_options = yes                   yes
+This is Ipopt version 3.12.8, running with linear solver mumps.
[…]
