### 2024-09-30 #7110 @Aravind-Sundararajan
> I did the following to further the review process.
>
>     1. I followed the instructions to create a new goal. I copied the MocoActivationGoal. I wanted to add the functionality of exponent of 3 and 4. I called it MocoCustomGoal. The directions are not very clear. It's not just the file names that have to be changed, but the names of classes and other variables might also need to be changed, depending on application. It also seems to me that the user needs to know how to use OpenSim API to actually make new custom goals. It seems that the user might need to have some knowledge of C++ as well. It would be nice to also add what qualifications or softwares the user needs to know to use MEP. Additionally, if the user already understands the OpenSim API, it should be added briefly how they benefit from MEP, that is, API does not provide this option.
>
>     2. In the test_ExtendProblem_simple.m, the opensimroot is hard coded. It needs to be user specified. The file also gives an error since it needs to be run from the top directory and not from inside the folder 'test'.  I think for OpenSim verion lower than 4.5, the _compat file is to be used. This is not mentioned anywhere. I was able to finally run the example and generate the figure that looks similar to the one in the readme section. Why is my figure not the same as the readme section (I have attached my figure here)? What do I need to do to replicate the result in ReadMe? My biggest problem with this example is that I have no idea what I ran. There is no file header for the user to understand what the code does. what does this example do? what are the task constraints? what is the goal here? what does the max coordinate goal do?
>        ![image](https://private-user-images.githubusercontent.com/85564166/369887539-1e0d969b-bf1b-4d04-80ff-cb7c1f287018.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3Mjc3MzE1NTUsIm5iZiI6MTcyNzczMTI1NSwicGF0aCI6Ii84NTU2NDE2Ni8zNjk4ODc1MzktMWUwZDk2OWItYmYxYi00ZDA0LTgwZmYtY2I3YzFmMjg3MDE4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MzAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTMwVDIxMjA1NVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTg1YmY5MWIxYjhjYmRmYzA2MjhmZjIwMjBkNDRhNTIxN2MwZjY2ZDdhMDM5MmNjYjU2NDQ3MDgxMzZjOTg1NDYmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.nb-Pe-N6AOu0H0iJR3BGtaDe5h536LRuawtckko-Qfo)
>
>     3. I ran the driver.m but again the filepaths are incorrect. The files need to be run form the top directory. If they need to be run from the top directory, why are they placed in the test directory? Which custom goals of MEP do the Walksim_Tracking.m and WalkSim_predictive.m use? What is the advantage, that is, what makes these simulations unique such that they cannot be used with regular Moco? I think that these information will help highlight the power of MEP. I was looking for the MocoActivationGoal to be used in the walking simulations but I only saw the MocoControlGoal, which is already a part of Moco, I want to use the new goal I created (in point 1) where I would penalize the activation^3 and activation^4. I thought that I would replace MocoActivationGoal with my custom goal to test the MEP. Could you add that possibility? I would also like more explanation of the tracking and predictive simulations of walking. Is it 2D or 3D? The header of the WalkSim_predictive_compat.m says that it is a tracking problem. What is the cost function here?
>
>     4. When I ran driver.m, the tracking part failed. I got an error saying "Error using WalkSim_Tracking
>        tracking failed to match reference output for goal". Why would that happen? An example code should not fail. What do I do to fix this? Also please provide reference results, that is, what should the results of running driver.m look like (for both tracking and predictive simulation) so that the user can be satisfied that they have installed everything correctly.
>
>     5. I think that if more documentation can be added to each of the new goals that are created in MEP, that will already be a great addition as people can just use those goals in their Moco. For example, I was excited to see the new Goals that are there in the custom_goals_compat folder, but I have no idea what each of them do and how to use them.
>
>
> I think that overall the MEP is very useful and can be highly impactful but it needs far better documentation for people to be able to use it.

I've hopefully addressed most of the issues in the document and added a tutorial section on goal creation. As for automating new goal creation, IMO this is beyond the scope of the project and copy-paste and a regex replace or Preserve Case replace in vscode is pretty simple enough.

As for tests sequestered in the test directory, I opt to add tests to the matlab path and this is a convention I use that is not abnormal; However, `MEP` can be used from anywhere if you add MocoExtendProblem/bin/RelWithDebInfo to the matlab path.
I instead opt to `addpath(genpath(fullfile(pwd,'bin','RelWithDebInfo')))` while staying on the top-level directory. Just a manner of taste.

And now let me address some of the issues with the examples/test files
