### 2024-09-30 95de7e8
fix: added compat output reference for 4.3 point mass
