### 2024-10-01 4cfc028
chore: readme including explanation of included goals
