### 2024-10-01 4d19f02
Update README.md explanation of tests
