### 2024-10-09 #6844 @aasadi1
I am following the tutorial and right now I'm running test_ExtendProblem_simple.m and I have a recommendation. Add a note in the documentation to change the openSim version in line 1, I changed mine to 4.4 so it ran. But now I get this error:

Error using createPointMass
Java exception occurred:
java.lang.RuntimeException: SimTK Exception thrown at Xml.cpp:115:
  Error detected by Simbody method Xml::writeToFile(): Failed to write to the Xml file
  './models/pointmass.osim' with error ' (line=1, col=1)'.
  (Required condition 'saveOK' was not met.)

	at org.opensim.modeling.opensimCommonJNI.OpenSimObject_print(Native Method)

	at org.opensim.modeling.OpenSimObject.print(OpenSimObject.java:603)

Error in test_ExtendProblem_simple (line 24)
p = createPointMass('./models/pointmass.osim', opensimroot);
