### 2024-10-09 #6844 @Aravind-Sundararajan
> I am following the tutorial and right now I'm running test_ExtendProblem_simple.m and I have a recommendation. Add a note in the documentation to change the openSim version in line 1, I changed mine to 4.4 so it ran. But now I get this error:
>
> ```
> Error using createPointMass
> Java exception occurred:
> java.lang.RuntimeException: SimTK Exception thrown at Xml.cpp:115:
>   Error detected by Simbody method Xml::writeToFile(): Failed to write to the Xml file
>   './models/pointmass.osim' with error ' (line=1, col=1)'.
>   (Required condition 'saveOK' was not met.)
>
>
> 	at org.opensim.modeling.opensimCommonJNI.OpenSimObject_print(Native Method)
>
> 	at org.opensim.modeling.OpenSimObject.print(OpenSimObject.java:603)
>
>
> Error in test_ExtendProblem_simple (line 24)
> p = createPointMass('./models/pointmass.osim', opensimroot);
> ```

I added a test_ExtendProblem_simple_compat.m where you will have to change the opensimroot to 'C:\opensim 4.4\'. After setting matlab to use my OpenSim 4.4 and rebuilding MocoExtendProblem, I was able to run test_ExtendProblem_simple_compat without issue.

It may be unable to write to the osim file because it is being used/opened by another program. Let me know if this is a sticking point and we can dig into it further.
