### 2024-10-20 #7110 @aasadi1
I finished my review. I was able to run the test without any errors. I did not edit any custom goals as I believe they need C++ knowledge. Last part says that:

"Now feel free to modify MocoCustomGoal.cpp and MocoCustomGoal.h for new goals! For help reference the [OpenSim API reference documentation](https://simtk.org/api_docs/opensim/api_docs/index.html)."

So we need toedit the .cpp or .h files to adjust the goal right? I was under the impression that this project is basically for those that only know MATLAB. I'm not sure someone like me with only MATLAB knowledge can adjust the .cpp and .h files to create custom goals. If it's possible, it needs more documentation on how to do that.

The paper looks really good and informative. Great job on it.

Let me know if I need to do anything else.
