### 2024-11-20 #7110 @dgupta7
It sounds like Azarang was already able to run the tests. So I am glad to hear that. I was also able to run the test_extendProblem_simple.m. Here are some comments:
1. I would like each test to start with a header that describes what the code does. I still do not know what the test_extendProblem_simple.m actually does.
2. Can you please add a description somewhere as to what each of the goals in the custom_goals folder do?
3. WalkSim_predictive takes an input sim_type. What value do I need to put in there? Please add that to the function header.
4. When I ran WalkSim_Tracking(), I still get the error "tracking failed to match reference output for goal". Am I correct in saying that this code is not using any MEP custom goals? All of it is commented out. If yes, why is it in this repository?
5. I was able to create a custom goal, but I haven't tested it yet.
