### 2024-12-20 9cd6a79
chore: added descriptions for test scripts
