### 2025-01-01 feb11fc
improved headers for gait simulations
