### 2025-01-01 cf832c2
updated compat gait script for parity with 4.5 script
