### 2025-01-01 #7110 @Aravind-Sundararajan
@aasadi1 @dgupta7 I have updated the scripts to reflect your suggestions and have tests correctly outputting when the simulation matches the result in `output_reference`, I assume there are no further comments in the scripts, README or paper. @Kevin-Mattheus-Moerman what can we do to push this review along?
