### 2025-01-01 897bb1d
corrected incorrect methodname for generate_content, removed erroneous utility script mot2MocoTrajectory
