+++ utils/generate_content.m
@@ -1,5 +1,5 @@
 function [cpp_out,mex_out,wrap_out] = generate_content(goal_tree)
-% generate_content_compat  Procedurally constructs the matlab wrapper, the
+% generate_content  Procedurally constructs the matlab wrapper, the
 % mex interface and the cpp class.
 %   [cpp_out,mex_out,wrap_out] = generate_content_compat(goal_tree)
 %
