### 2025-01-09 a694db3
chore: updating paper for clarity
