+++ paper/paper.md
@@ -48,9 +48,9 @@ This repository features:
 
 OpenSim is an open-source software platform for modeling musculoskeletal structures and creating dynamic simulations of movement [@delp2007,@Seth2018]. OpenSim enables researchers and clinicians to investigate how biological and non-biological structures respond to different loads, postures and activities in both static and dynamic situations. OpenSim has been used to study a wide range of biomechanical problems, such as the mechanics of walking and running [e.g. @Falisse2019], the impact of injury or disease on movement [e.g. @Johnson2022], and the effectiveness of rehabilitation exercises [e.g. @Spomer2023].
  
-While OpenSim originally featured several single-shooting methods like Computed Muscle Control and static optimization to solve for kinematics, kinetics and controls, OpenSim now includes Moco [@Dembia2020] which employs an optimization paradigm called direct collocation for solving curve-fitting problems that range from solving for muscle forces, to tracking experimental data, and fully predictive simulations that solve the optimal control problem (OCP) subject to some predefined goals and constraints. Direct collocation is a numerical optimal control method [@Kelly2017] that is computationally efficient compared to single-shooting algorithms and is used extensively in computational approaches to understanding biological movement. While direct collocation is powerful, OpenSim Moco only provides a small predefined set of optimization goals which can be modified easily using OpenSim's MATLAB API; However nmore sophisticated goals such as the 3 stability criteria explored in the showcases require understanding the C++ API and ability to build a custom plugin or building OpenSim from source. It can be daunting for many users to develop custom goals without experience building software written in C++. We developed `MEP` so Moco users without experience compiling C++ can still write and test custom goals.
+While OpenSim originally featured several single-shooting methods like Computed Muscle Control and static optimization to solve for kinematics, kinetics and controls, OpenSim now includes Moco [@Dembia2020] which employs an optimization paradigm called direct collocation for solving curve-fitting problems that range from solving for muscle forces, to tracking experimental data, and fully predictive simulations that solve the optimal control problem (OCP) subject to some predefined goals and constraints. Direct collocation is a numerical optimal control method [@Kelly2017] that is computationally efficient compared to single-shooting algorithms and is used extensively in computational approaches to understanding biological movement. While direct collocation is powerful, OpenSim Moco only provides a small predefined set of optimization goals which can be modified easily using OpenSim's MATLAB API; However nmore sophisticated goals such as the 3 stability criteria explored in the showcases require understanding the C++ API and ability to build a custom plugin or building OpenSim from source. It can be daunting for many users to develop custom goals without experience in building software written in C++. We developed `MEP` so Moco users without experience in compiling C++ can still write and test custom goals.
 
-`MEP` was developed using MATLAB (v. 2022a), which is a multimodal software platform that is commonly used by biomechanics researchers. Typically, OpenSim interfaces are generated automatically with SWIG (Simplified Wrapper and Interface Generator), as opposed to developing an interface with MATLAB classes and MEX (MATLAB Executable), which can be challenging for even experienced biomechanists. `MEP` only requires that CMake and msbuild from Visual Studio (VS) 2019 or higher as well as the C++ desktop development workload for VS to use MATLAB’s MEX compiler with VS.
[…]
