### 2025-01-09 bf90c59
fix: made opensim dir selection more flexible
