### 2025-01-09 9364705
chore: making test scripts more flexible with opensim version
