### 2025-01-09 cd4da6a
fix: parity in simple compat test for addpaths
