+++ test/test_ExtendProblem_simple_compat.m
@@ -55,6 +55,7 @@ import org.opensim.modeling.*
 
 
 addpath(genpath(fullfile(pwd,'utils'))); %utilities
+addpath(genpath(fullfile(pwd,'test'))); %test
 addpath(genpath(fullfile(pwd,'models'))); %utilities
 addpath(genpath(fullfile(pwd,'sandbox'))); %sandbox
 addpath(genpath(fullfile(pwd,'bin','RelWithDebInfo'))); %Extend Problem (magic!)
