### 2025-01-12 Aravind-Sundararajan/MocoExtendProblem#6 @0todd0000
[build error]  MSBUILD : error MSB1009: Project file does not exist
I'm trying to follow the `Compile MEX interface` instructions in Getting Started, but encounter the following error:

```
MSBUILD : error MSB1009: Project file does not exist.
Switch: C:\Temp\MocoExtendProblem-main\build\customGoals.sln
detected using opensim 4.5.
Error using fprintf
Invalid file identifier. Use fopen to generate a valid file identifier.

Error in build_extend_class (line 29)
fprintf(fid,'%s\n',s);

Error in build (line 72)
build_extend_class(fullfile(bindir,config,cppName),fullfile(bindir,config,wrapName), opensim_install);
```

I think this is just a path problem but I can't seem to find how to fix this error.

Note that `C:\OpenSim 4.5\bin` is indeed on my PATH, but that the `getenv` command on Line 16 of build.m returns empty:

```matlab
getenv('OPENSIM_HOME')
```

and that `exist('opensim_config.mat', 'file')` on Line 18 is True.

Please help!
