### 2025-01-12 #7110 @0todd0000
Hello and apologies for the delay. I'm still encountering getting started difficulties and have just posted a new issue [here](https://github.com/Aravind-Sundararajan/MocoExtendProblem/issues/6)
