### 2025-01-12 Aravind-Sundararajan/MocoExtendProblem#6 @Aravind-Sundararajan
@0todd0000

it should be okay that `OPENSIM_HOME` isn't assigned and `opensim_config.mat` had been generated. I would verify that the `opensim_install` variable in `opensim_config.mat` is a path like `C:\OpenSim 4.5\` without the bin folder. if cmake and msbuild ran correctly, there should be a `customGoals.sln` in the build folder of MocoExtendProblem but it looks like there was an issue with cmake or msbuild. I would make sure build.m is being run with matlab's current folder as the top-level directory of MocoExtendProblem.
