### 2025-01-12 d1018fd
chore: refactor made opensim version more flexible
