### 2025-01-13 Aravind-Sundararajan/MocoExtendProblem#6 @Aravind-Sundararajan
> * `opensim_install` in `opensim_config.mat` is indeed `'C:\OpenSim 4.5\'`
>
>     * There is no `customGoals.sln` in the build folder.
>
>
> > I would make sure build.m is being run with matlab's current folder as the top-level directory of MocoExtendProblem.
>
> Yes, here is what I am running:
>
> ```matlab
> cd("C:\Temp\MocoExtendProblem-main")
> build
> ```
>
> Here is the full error:
>
> ```
> CMake Deprecation Warning at CMakeLists.txt:1 (cmake_minimum_required):
>   Compatibility with CMake < 3.10 will be removed from a future version of
>   CMake.
>
>   Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
>   to tell CMake that the project requires at least <min> but has been updated
>   to work with policies introduced by <max> or earlier.
>
>
> CMake Error at CMakeLists.txt:2 (project):
>   Generator
>
>     Visual Studio 17 2022
>
>   could not find any instance of Visual Studio.
>
>
>
> -- Configuring incomplete, errors occurred!
> Microsoft (R) Build Engine version 4.8.9032.0
> [Microsoft .NET Framework, version 4.0.30319.42000]
> Copyright (C) Microsoft Corporation. All rights reserved.
>
> MSBUILD : error MSB1009: Project file does not exist.
> Switch: C:\Temp\MocoExtendProblem-main\build\customGoals.sln
> detected using opensim 4.5.
> Error using fprintf
> Invalid file identifier. Use fopen to generate a valid file identifier.
>
> Error in build_extend_class (line 29)
> fprintf(fid,'%s\n',s);
>
> Error in build (line 72)
> build_extend_class(fullfile(bindir,config,cppName),fullfile(bindir,config,wrapName), opensim_install);
> ```
>
> It looks like Visual Studio is not being found? I have Visual Studio 2022 installed.

You may be missing the Visual Studio C++ tools for cmake feature or the entire c++ for desktop development workflow. you can reach the installer from visual studio with `Tools > Get Tools and Features`

here are the optional components I have installed
![image](https://github.com/user-attachments/assets/8dff6968-b620-4b15-9ea2-067608fed7b4)

If you are missing anything, I would modify install and let visual studio restart your computer if it asks you to.
