### 2025-01-14 Aravind-Sundararajan/MocoExtendProblem#6 @Aravind-Sundararajan
> OK, thanks for this. Installing the C++ tools allowed me to go forward, albeit with the following error:
>
> ```
> CMake Deprecation Warning at CMakeLists.txt:1 (cmake_minimum_required):
>   Compatibility with CMake < 3.10 will be removed from a future version of
>   CMake.
>
>   Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
>   to tell CMake that the project requires at least <min> but has been updated
>   to work with policies introduced by <max> or earlier.
>
>
> -- Selecting Windows SDK version 10.0.22621.0 to target Windows 10.0.26100.
> -- The C compiler identification is MSVC 19.42.34435.0
> -- The CXX compiler identification is MSVC 19.42.34435.0
> -- Detecting C compiler ABI info
> -- Detecting C compiler ABI info - done
> -- Check for working C compiler: C:/Program Files/Microsoft Visual Studio/2022/Community/VC/Tools/MSVC/14.42.34433/bin/Hostarm64/arm64/cl.exe - skipped
> -- Detecting C compile features
> -- Detecting C compile features - done
> -- Detecting CXX compiler ABI info
> -- Detecting CXX compiler ABI info - done
> -- Check for working CXX compiler: C:/Program Files/Microsoft Visual Studio/2022/Community/VC/Tools/MSVC/14.42.34433/bin/Hostarm64/arm64/cl.exe - skipped
> -- Detecting CXX compile features
> -- Detecting CXX compile features - done
> OSim_Version=5
> Source Directory=C:/Temp/MocoExtendProblem-main/custom_goals/
> -- Performing Test CMAKE_HAVE_LIBC_PTHREAD
> -- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Failed
> -- Looking for pthread_create in pthreads
> -- Looking for pthread_create in pthreads - not found
> -- Looking for pthread_create in pthread
> -- Looking for pthread_create in pthread - not found
> -- Found Threads: TRUE
> looking for custom goals:
> osimMocoActivationGoal:
> osimMocoActivationSquaredGoal:
> osimMocoBOSGoal:
> osimMocoCoordinateAccelerationGoal:
> osimMocoMarkerAccelerationGoal:
> osimMocoMaxCoordinateGoal:
> osimMocoZMPGoal:
> test:testCustomGoal.exe
> -- Configuring done (16.0s)
> -- Generating done (0.5s)
> -- Build files have been written to: C:/Temp/MocoExtendProblem-main/build
> Microsoft (R) Build Engine version 4.8.9032.0
> [Microsoft .NET Framework, version 4.0.30319.42000]
> Copyright (C) Microsoft Corporation. All rights reserved.
>
> Building the projects in this solution one at a time. To enable parallel build, please add the "/m" switch.
> Build started 2025-01-14 10:38:50 AM.
> Project "C:\Temp\MocoExtendProblem-main\build\customGoals.sln" on node 1 (default targets).
> ValidateSolutionConfiguration:
>   Building solution configuration "RelWithDebInfo|ARM64".
> ValidateProjects:
>   The project "INSTALL" is not selected for building in solution configuration "RelWithDebInfo|ARM64".
> Project "C:\Temp\MocoExtendProblem-main\build\customGoals.sln" (1) is building "C:\Temp\MocoExtendProblem-main\build\ALL_BUILD.vcxproj.metaproj" (2) on node 1 (default targets).
> Project "C:\Temp\MocoExtendProblem-main\build\ALL_BUILD.vcxproj.metaproj" (2) is building "C:\Temp\MocoExtendProblem-main\build\Copy_OpenSim_DLLs.vcxproj.metaproj" (3) on node 1 (default targets).
> Project "C:\Temp\MocoExtendProblem-main\build\Copy_OpenSim_DLLs.vcxproj.metaproj" (3) is building "C:\Temp\MocoExtendProblem-main\build\ZERO_CHECK.vcxproj" (4) on node 1 (default targets).
> C:\Temp\MocoExtendProblem-main\build\ZERO_CHECK.vcxproj(36,3): error MSB4019: The imported project "C:\Microsoft.Cpp.Default.props" was not found. Confirm that the path in the <Import> declaration is correct, and that the file exists on disk.
> Done Building Project "C:\Temp\MocoExtendProblem-main\build\ZERO_CHECK.vcxproj" (default targets) -- FAILED.
> Done Building Project "C:\Temp\MocoExtendProblem-main\build\Copy_OpenSim_DLLs.vcxproj.metaproj" (default targets) -- FAILED.
> Done Building Project "C:\Temp\MocoExtendProblem-main\build\ALL_BUILD.vcxproj.metaproj" (default targets) -- FAILED.
> Done Building Project "C:\Temp\MocoExtendProblem-main\build\customGoals.sln" (default targets) -- FAILED.
>
> Build FAILED.
>
> "C:\Temp\MocoExtendProblem-main\build\customGoals.sln" (default target) (1) ->
> "C:\Temp\MocoExtendProblem-main\build\ALL_BUILD.vcxproj.metaproj" (default target) (2) ->
> "C:\Temp\MocoExtendProblem-main\build\Copy_OpenSim_DLLs.vcxproj.metaproj" (default target) (3) ->
> "C:\Temp\MocoExtendProblem-main\build\ZERO_CHECK.vcxproj" (default target) (4) ->
>   C:\Temp\MocoExtendProblem-main\build\ZERO_CHECK.vcxproj(36,3): error MSB4019: The imported project "C:\Microsoft.Cpp.Default.props" was not found. Confirm that the path in the <Import> declaration is correct, and that the file exists on disk.
>
>     0 Warning(s)
>     1 Error(s)
>
> Time Elapsed 00:00:02.64
> detected using opensim 4.5.
> Error using fprintf
> Invalid file identifier. Use fopen to generate a valid file identifier.
>
> Error in build_extend_class (line 29)
> fprintf(fid,'%s\n',s);
>
> Error in build (line 72)
> build_extend_class(fullfile(bindir,config,cppName),fullfile(bindir,config,wrapName), opensim_install);
> ```
>
> `customGoals.sln` now successfully appears in the build directory along with several other folders and files.
>
> Neither installing .NET desktop development nor restarting changed the error above.
>
> Not sure what to do next, please help!

it's possible this is a PATH problem, for reference my system environment PATH has the following
![image](https://github.com/user-attachments/assets/1541bdec-cb80-4bd4-bcb3-6230924bb548)
so I have paths for msbuild from .net and visual studio 2022 in my system PATH and I can build fine.

I would try cleaning/deleting the build folder, it is possible that something is getting cached.

it's possible the environment variable `VCTargetsPath` isn't getting set, for which you want the path `C:\Program Files\Microsoft Visual Studio\2022\Community\MSBuild\Microsoft\VC\v170\`
This directory has the missing `Microsoft.Cpp.Default.props`
