### 2025-01-14 Aravind-Sundararajan/MocoExtendProblem#6 @0todd0000
> it's possible this is a PATH problem, for reference my system environment PATH has the following.. so I have paths for msbuild from .net and visual studio 2022 in my system PATH and I can build fine.

I had the Microsoft.NET directory on the system path but not the Visual Studio 2022 Bin directory.  I added this VS 2022 Bin directory, but unfortunately this did not solve things. Still I'm getting the same error as above.

> I would try cleaning/deleting the build folder, it is possible that something is getting cached.

Agreed, I've been deleting everything before each build attempt.

> it's possible the environment variable VCTargetsPath isn't getting set, for which you want the path C:\Program Files\Microsoft Visual Studio\2022\Community\MSBuild\Microsoft\VC\v170\
This directory has the missing Microsoft.Cpp.Default.props

I tried manually adding `VCTargetsPath` to the system environment variables as suggested, restarting Matlab and cleaning the build directory.  Unfortunately the same error occurs.

I don't suppose that the build problems are caused by virtualized Windows?  I am running Windows 11 via Parallels on my Mac.  I unfortunately do not have easy access to a proper Windows PC.  The reason I ask is that I noticed `RelWithDebInfo|ARM64` in the error message, but my Mac has an Apple M3 processor.  I originally tried building on Mac ( #5 )  but just realized that I failed to mention in this thread that I'm now trying to build on virtualized Windows.
