### 2025-01-14 Aravind-Sundararajan/MocoExtendProblem#6 @Aravind-Sundararajan
> > it's possible this is a PATH problem, for reference my system environment PATH has the following.. so I have paths for msbuild from .net and visual studio 2022 in my system PATH and I can build fine.
>
> I had the Microsoft.NET directory on the system path but not the Visual Studio 2022 Bin directory.  I added this VS 2022 Bin directory, but unfortunately this did not solve things. Still I'm getting the same error as above.
>
>
> <br>
> <br>
>
> > I would try cleaning/deleting the build folder, it is possible that something is getting cached.
>
> Agreed, I've been deleting everything before each build attempt.
>
> <br>
> <br>
>
> > it's possible the environment variable VCTargetsPath isn't getting set, for which you want the path C:\Program Files\Microsoft Visual Studio\2022\Community\MSBuild\Microsoft\VC\v170\
> This directory has the missing Microsoft.Cpp.Default.props
>
> I tried manually adding `VCTargetsPath` to the system environment variables as suggested, restarting Matlab and cleaning the build directory.  Unfortunately the same error occurs.
>
> <br>
> <br>
>
> I don't suppose that the build problems are caused by virtualized Windows?  I am running Windows 11 via Parallels on my Mac.  I unfortunately do not have easy access to a proper Windows PC.  The reason I ask is that I noticed `RelWithDebInfo|ARM64` in the error message, but my Mac has an Apple M3 processor.  I originally tried building on Mac ( #5 )  but just realized that I failed to mention in this thread that I'm now trying to build on virtualized Windows.

I'm not too sure, it may be impossible but it works up to here so I would be surprised if it didn't actually work on virtualized windows.

In command prompt, if you use `where msbuild`, what path does it say it's using? Could also try opening the .sln file in visual studio itself and build the project by hand though I haven't even tried that on windows.

It may just not work on parallels though I can't be certain.
