### 2025-01-15 Aravind-Sundararajan/MocoExtendProblem#6 @0todd0000
I just realized that this discussion has drifted from the thread title, and that the initial problem has been solved. Should this issue be closed and a new one opened?
