### 2025-02-06 #7110 @aasadi1
@Aravind-Sundararajan The only parts missing from my review are in the documentation section and are: functionality documentation, automated tests, community guidelines. Here's my comments about these parts:

Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I think for this part you need to add more details. I know that this software supposed to help with creating custom moco goals in MATLAB but I am still confused how I can achieve that. My understanding is that you need to create the goal in C++ and then call it into your MATLAB script, right? I need some more clarification in this part.

Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I will try to run the test in your paper to see if I can generate your results this weekend.

Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

For this part, I didn't see anything. Please direct me toward the documentation for this part.

I'll update the checklist this weekend after running the automated tests.
