### 2025-02-07 0f79cad
Update README.md
clarifying summary so it is more clear about C++
