### 2025-02-07 c96d1f4
chore: added info about api documentation
