### 2025-02-07 #7110 @Aravind-Sundararajan
Hi @aasadi1, to answer your questions

>My understanding is that you need to create the goal in C++ and then call it into your MATLAB script, right? I need some more clarification in this part.

You are correct that you need to design and write your custom goal in C++, use the `build.m` script to create the MEX and the procedurally constructed ExtendProblem class. I updated the summary in the README to clarify this. The [API documentation](https://aravind-sundararajan.github.io/MocoExtendProblem) is also provided.

>automated tests

I did not add any automated testing, but there are tests provided that the user can run as described in the README. Though I think I could with github actions, I would have to install opensim on the runner in the workflow and I thought that was overkill and complicated, it should be sufficient for the user to run the simple and gait tests manually I think.

>community guidelines

In the repository  there are are a [code of conduct](https://github.com/Aravind-Sundararajan/MocoExtendProblem/blob/main/CODE_OF_CONDUCT.md) and [contributing](https://github.com/Aravind-Sundararajan/MocoExtendProblem/blob/main/CONTRIBUTING.md) that should suffice, let me know if this is okay.
