### 2025-02-07 b18f90b
chore: Update paper.md so it is more clear that goals are written in C++
