### 2025-02-07 4502c8e
Update main.yml to correct old upload-artifact version
