+++ .github/workflows/main.yml
@@ -3,7 +3,7 @@ on:
   push:
     paths:
       - paper/**
-      - .github/workflows/draft-pdf.yml
+      - .github/workflows/main.yml
       - paper/file_tree.PNG
       - paper/stability.PNG
 jobs:
