### 2025-02-13 #7110 @0todd0000
@Kevin-Mattheus-Moerman
Please advise re: my post last week, copied here:

> [@Kevin-Mattheus-Moerman](https://github.com/Kevin-Mattheus-Moerman) I'm stuck on building for virtual Windows as detailed in [Issue #6](https://github.com/Aravind-Sundararajan/MocoExtendProblem/issues/6) in the project repository. I thought I could access a Windows machine but grad students need to use my best Windows option for their simulation work so I'm a bit stuck in terms of hardware at the moment. My opinion of this project is high in terms of its purpose, scientific utility, documentation and support, so I'd be pleased to recommend publication based on these points alone, especially given the authors' rapid and helpful Issues responses. I am simply struggling to build it so I can't yet tick all the reviewer boxes.
>
> Is it possible to recommend publication without having ticked all reviewer checkboxes?
>
> As you will see in my checklist the boxes I have not yet checked are:
>
> * Reproducibility
> * Installation
> * Functionality
> * Performance
> * Automated tests
