### 2025-02-23 #7110 @dgupta7
I think we are close to the finish line.

I ran the test_ExtendProblem_simple.m with the line ep.addMocoActivationSquaredGoal('act_square',1.0, true, 0.); uncommented. It gave me an error of not enough input arguments:

Not enough input
arguments.

Error in extend_problem/addMocoActivationSquaredGoal (line 25)
            extendProblem('addMocoActivationSquaredGoal', this.id_,goalName,weight, div_disp, div_dur, div_mass, end_point_goal);

Error in test_ExtendProblem_simple (line 120)
        ep.addMocoActivationSquaredGoal('act_square',1.0, true, 0.);
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
Could you please fix it. As my final test, I want to change the exponent and test how the results change.
