### 2025-02-23 6c62011
fix: incorrect arguments in test simple for ep method
