### 2025-02-23 56a45a4
made simple tests more flexible with trying commented alternate goals
