### 2025-02-23 #7110 @dgupta7
Thanks for the update. When I ran the test with and without that goal, the results did not change at all. Does this problem actually use the activationsquaredgoal? Or maybe could I test this goal in a walking simulations?
