### 2025-02-23 #7110 @Aravind-Sundararajan
> > Thanks for the update. When I ran the test with and without that goal, the results did not change at all. Does this problem actually use the activationsquaredgoal? Or maybe could I test this goal in a walking simulations?
>
> [@dgupta7](https://github.com/dgupta7) the goal should work if you used the activation goal, though it may not change the kinematics perceptibly... you could check the activations or excitations in the output to see if they are different from the simple test w/o any extend_problem goal.
>
> The other goals there that are commented out have much more obvious changes to the kinematics of the simple model. You could try using a more complex model like the gait model or changing the constrants in the simple model study to see if you get a change.

[`m_power_function` is set here](https://github.com/Aravind-Sundararajan/MocoExtendProblem/blob/56a45a4d29633b4efb3aded4e14030f80f53ee1e/custom_goals/MocoActivationSquaredGoal/MocoActivationSquaredGoal.cpp#L37-L49) is intended to swap in for `SimTK:square` for different exponents in  [`calcIntegrandImpl()`](https://github.com/Aravind-Sundararajan/MocoExtendProblem/blob/56a45a4d29633b4efb3aded4e14030f80f53ee1e/custom_goals/MocoActivationSquaredGoal/MocoActivationSquaredGoal.cpp#L37-L49)

for an example of using m_power_function, here the strategy is being used in opensim core's [MocoControlGoal](https://github.com/opensim-org/opensim-core/blob/3aaae0035e9fac62be4902b7f62f16b300134bc9/OpenSim/Moco/MocoGoal/MocoControlGoal.cpp#L168-L188).
