### 2025-02-24 #7110 @dgupta7
I don't think the goal is working. Here is what I did:
1. I ran test_ExtendProblem_simple.m as it is. The solution is sliding_mass_solution.sto.

2. I uncommented the following lines and ran it again. The solution is called sliding_mass_solution_activationSquaredGoal.stocustom_goal_name = 'act_square';
ep.addMocoActivationSquaredGoal(custom_goal_name, 1.0, false, false, false, false);

3. I was not sure if I am implementing the goal correctly since I could not find any documentation of what each of the 6 inputs mean. So using my best judgement, I made the following change and the solution is called sliding_mass_solution_activationSquaredGoal_actuator.sto.
ep.addMocoActivationSquaredGoal(custom_goal_name, 1.0, false, false, false, '/actuator');

4. I thought that maybe the weight is too low, so I changed the goal to the following. The solution is called sliding_mass_solution_activationSquaredGoal_actuator_w1000000.sto.
ep.addMocoActivationSquaredGoal(custom_goal_name, 1000000.0, false, false, false, '/actuator');

All of the 4 solutions I got were exactly the same. Now, it is possible that I might be using the goal incorrectly or I might be using the wrong inputs. Could you please direct me to the documentation where the meaning of each of the inputs is defined and add that information in the readme? Thanks!

PS: Github was not allowing me to upload the .sto files, to I changed them to .txt before uploading

[sliding_mass_solution_activationSquaredGoal.txt](https://github.com/user-attachments/files/18944695/sliding_mass_solution_activationSquaredGoal.txt)
[sliding_mass_solution_activationSquaredGoal_actuator.txt](https://github.com/user-attachments/files/18944696/sliding_mass_solution_activationSquaredGoal_actuator.txt)
[sliding_mass_solution_activationSquaredGoal_actuator_w1000000.txt](https://github.com/user-attachments/files/18944698/sliding_mass_solution_activationSquaredGoal_actuator_w1000000.txt)
[sliding_mass_solution.txt](https://github.com/user-attachments/files/18944697/sliding_mass_solution.txt)
