### 2025-02-25 7e9b48c
corrected activation goals constructProperty missing for exponent
