### 2025-02-25 #7110 @Aravind-Sundararajan
> I don't think the goal is working. Here is what I did:
>
>     1. I ran test_ExtendProblem_simple.m as it is. The solution is sliding_mass_solution.sto.
>
>     2. I uncommented the following lines and ran it again. The solution is called sliding_mass_solution_activationSquaredGoal.sto
>        custom_goal_name = 'act_square';
>        ep.addMocoActivationSquaredGoal(custom_goal_name, 1.0, false, false, false, false);
>
>     3. I was not sure if I am implementing the goal correctly since I could not find any documentation of what each of the 6 inputs mean. So using my best judgement, I made the following change and the solution is called sliding_mass_solution_activationSquaredGoal_actuator.sto.
>        ep.addMocoActivationSquaredGoal(custom_goal_name, 1.0, false, false, false, '/actuator');
>
>     4. I thought that maybe the weight is too low, so I changed the goal to the following. The solution is called sliding_mass_solution_activationSquaredGoal_actuator_w1000000.sto.
>        ep.addMocoActivationSquaredGoal(custom_goal_name, 1000000.0, false, false, false, '/actuator');
>
>
> All of the 4 solutions I got were exactly the same. Now, it is possible that I might be using the goal incorrectly or I might be using the wrong inputs. Could you please direct me to the documentation where the meaning of each of the inputs is defined and add that information in the readme? Thanks!
>
> PS: Github was not allowing me to upload the .sto files, to I changed them to .txt before uploading
>
> [sliding_mass_solution_activationSquaredGoal.txt](https://github.com/user-attachments/files/18944695/sliding_mass_solution_activationSquaredGoal.txt) [sliding_mass_solution_activationSquaredGoal_actuator.txt](https://github.com/user-attachments/files/18944696/sliding_mass_solution_activationSquaredGoal_actuator.txt) [sliding_mass_solution_activationSquaredGoal_actuator_w1000000.txt](https://github.com/user-attachments/files/18944698/sliding_mass_solution_activationSquaredGoal_actuator_w1000000.txt) [sliding_mass_solution.txt](https://github.com/user-attachments/files/18944697/sliding_mass_solution.txt)

Hello @dgupta7 , I forgot, if you look at the implementation of MocoActivationSquaredGoal, it only adds to the cost term if the actuator in the model is a **muscle**, your solutions should all be the same and your cost term should be 0.

I'm on 4.4 and I can confirm that I can add the lines:

```
cptr = uint64(problem.getCPtr(problem));
ep = extend_problem(cptr);
ep.addMocoActivationSquaredGoal('act_square', 1.0, true, 0.);
```
to either the tracking or predictive human gait model simulations and it will run, giving you an "objective_control_effort and "objective_act_squared" in your solution file. Let me know if you can use these in the version you are using.

For the arguments, I can amend the README so these are better explained, but you can look at the `bin/RelWithDebInfo/extendProblem.cpp` and  `bin/RelWithDebInfo/extend_problem.m` to see what arguments are required from the procedurally generated methods/bindings.

- MocoActivationSquaredGoal
  * a slight modification to the MocoActivationGoal that squares the activations, note this is still different from the MocoControlEffortGoal that is built into Moco

in opensim 4.5:

    `void 	addMocoActivationSquaredGoal (const std::string &goalName, double weight, bool div_disp, bool div_dur, bool div_mass, double end_point_goal)`

    `const std::string &goalName`: goal name

    `double weight` : goal weight

    `bool div_disp`: divide cost by displacement

    `bool div_dur`: divide cost by time duration

    `bool div_mass`: divide cost by mass

    `double end_point_goal`: endpoint constraint mode value
