### 2025-03-13 5f8754f
swapped 4.5 activation goal to use m_power_function by user request
