### 2025-03-13 4f00f6c
added get, set methods for exponent
