### 2025-03-15 505e714
activation goal now allows specifying individual weights
