+++ custom_goals_compat/MocoBOSGoal/MocoBOSGoal.h
@@ -2,77 +2,143 @@
 #define OPENSIM_MOCOBOSGOAL_H
 /* -------------------------------------------------------------------------- *
  * OpenSim: MocoBOSGoal.h                                                     *
- * -------------------------------------------------------------------------- *
- *                                                                            *
- * Author(s): Aravind Sundararajan, Varun Joshi                               *
  * -------------------------------------------------------------------------- */
-
 
 #include <OpenSim/Moco/osimMoco.h>
 #include "osimMocoBOSGoalDLL.h"
 
 namespace OpenSim {
 
-struct comData{
-  std::map<SimTK::MobilizedBodyIndex, SimTK::Matrix> Fs;  //!<@brief  support force (spatial (6x1))
-  std::map<SimTK::MobilizedBodyIndex, SimTK::Vec3> cop; //!<@brief  center of pressure (3x1)
-  std::map<SimTK::MobilizedBodyIndex, SimTK::SpatialVec> force; //!<@brief  linear summed cop force and resultant moment (3x2)
+/** @brief Structure containing data related to center of mass and support forces
+ *
+ * This structure stores maps that relate mobilized body indices to their
+ * respective support forces, centers of pressure, and resultant forces/moments.
+ */
+struct comData {
+    std::map<SimTK::MobilizedBodyIndex, SimTK::Matrix> Fs;     //!< Support force (spatial 6x1)
+    std::map<SimTK::MobilizedBodyIndex, SimTK::Vec3> cop;      //!< Center of pressure (3x1)
+    std::map<SimTK::MobilizedBodyIndex, SimTK::SpatialVec> force; //!< Linear summed COP force and resultant moment (3x2)
 };
 
+/** @brief Goal that minimizes the deviation of the base of support (BOS) from a target
+ *
+ * This goal minimizes the deviation of the model's base of support from a desired
+ * target location, integrated over the phase. The base of support is calculated
+ * using the contact forces and center of pressure between the feet and the ground.
+ *
+ * This goal can be useful for:
+ * - Improving balance and stability in walking simulations
+ * - Controlling foot placement and weight transfer
+ * - Minimizing risk of falls
+ *
+ * The goal requires specification of frames for both feet to properly calculate
+ * the base of support. The deviation can optionally be divided by the total 
+ * displacement of the model during the phase to make the cost invariant to the 
+ * distance traveled.
+ */
 class OSIMMOCOBOSGOAL_API MocoBOSGoal : public MocoGoal {
     OpenSim_DECLARE_CONCRETE_OBJECT(MocoBOSGoal, MocoGoal);
 
 public:
-    MocoBOSGoal() { constructProperties();}
+    /** @name Constructors */
+    /// @{
+    /** Default constructor */
+    MocoBOSGoal() { constructProperties(); }
+
+    /** Constructor with name
+     * @param name The name of the goal */
     MocoBOSGoal(std::string name) : MocoGoal(std::move(name)) {
         constructProperties();
     }
+
+    /** Constructor with name and weight
+     * @param name The name of the goal
+     * @param weight Weight for this goal term in the optimization */
     MocoBOSGoal(std::string name, double weight)
             : MocoGoal(std::move(name), weight) {
         constructProperties();
     }
-	
-    // Public members to change the divide by displacement property
+    /// @}
+
+    /** Set the exponent for the deviation term
+     * @param ex The exponent value */
     void setExponent(int ex) { set_exponent(ex); }
-    bool getExponent() const {
-        return get_exponent();
-    }
-	
-    // Public members to change the divide by displacement property
+    
+    /** Get the current exponent value
+     * @return The exponent value */
+    bool getExponent() const { return get_exponent(); }
+
+    /** Set whether to divide by displacement
[…]

[…]
