### 2025-03-15 ddd588a
added automated generation of documentation in procedurally generated extend_problem.m
