### 2025-03-15 d7dc94a
made the cmake more flexible and used opensim to get the version instead of the install directory
