+++ CMakeLists.txt
@@ -3,19 +3,13 @@ project(customGoals VERSION 1.0.2 DESCRIPTION "some custom goals")
 
 set(CMAKE_CXX_STANDARD 11)
 
-message("OSim_Version=${OSim_Version}")
-if( OSim_Version EQUAL 5) 
-    set(SRCDIR ${CMAKE_CURRENT_SOURCE_DIR}/custom_goals/)    
-    set(OSIMVersion "4.5")
-elseif(OSim_Version EQUAL 4)
+message("OSimVersion=${OSimVersion}")
+if(OSimVersion VERSION_GREATER_EQUAL "4.5")
+    set(SRCDIR ${CMAKE_CURRENT_SOURCE_DIR}/custom_goals/)
+    message(STATUS "Using OpenSim ${OSimVersion} - standard mode")
+else()
     set(SRCDIR ${CMAKE_CURRENT_SOURCE_DIR}/custom_goals_compat/)
-    set(OSIMVersion "4.4")
-elseif(OSim_Version EQUAL 3)
-    set(SRCDIR ${CMAKE_CURRENT_SOURCE_DIR}/custom_goals_compat/)
-    set(OSIMVersion "4.3")
-elseif(OSim_Version EQUAL 2)
-    set(SRCDIR ${CMAKE_CURRENT_SOURCE_DIR}/custom_goals_compat/)
-    set(OSIMVersion "4.2")
+    message(STATUS "Using OpenSim ${OSimVersion} - compatibility mode")
 endif()
 message("Source Directory=${SRCDIR}")
 
