### 2025-03-15 #7110 @Aravind-Sundararajan
@dgupta7 Some changes were made among which the build script will now generate documentation for each of the add goal methods of the wrapper class, which can be viewed by calling `help extend_problem` in the command window when the class is included in the MATLAB path. Hopefully this addresses the documentation concerns.

![Image](https://github.com/user-attachments/assets/b7413b94-0e01-4fc6-9aec-f25b43487ead)
