### 2025-03-15 170212e
updated upweighted muscles in effort predictive sim to right and left hamstrings
