### 2025-03-19 de34786
changed the MocoActivationGoal power 1 function from abs to sqrt(x^2 + e)
