+++ custom_goals/MocoActivationGoal/MocoActivationGoal.cpp
@@ -55,7 +55,7 @@ void MocoActivationGoal::initializeOnModelImpl(const Model& model) const {
       // The pow() function gives slightly different results than x * x. On Mac,
       // using x * x requires fewer solver iterations.
       if (exponent == 1) {
-        m_power_function = [](const double &x) { return std::abs(x); };
+        m_power_function = [](const double &x) { return std::sqrt(x*x + 1e-8); };
       } else if (exponent == 2) {
         m_power_function = [](const double &x) { return x * x; };
       } else {
