### 2025-03-23 cc687ce
cleaned up print statements in activation goal
