### 2025-03-23 b92ee68
changed the behavior of MocoCoordinateAccelerationGoal to match modifications to MocoActivationGoal
