+++ test/example2DWalking/WalkSim_predictive.m
@@ -196,9 +196,14 @@ if goal_type == 1
     musc_weights = [2000.0,2000.0];
     ep.addMocoActivationGoal(custom_goal_name, 1.0, false, false, false, false, 2,musc_names,musc_weights);
 elseif goal_type ==2
-    ep.addMocoMarkerAccelerationGoal('addMocoMarkerAccelerationGoal',10.0,...
-        div_disp, div_dur, div_mass,...
-        char(model.getMarkerSet().get("head_marker").getAbsolutePathString()));
+%     ep.addMocoMarkerAccelerationGoal('addMocoMarkerAccelerationGoal',10.0,...
+%         div_disp, div_dur, div_mass,...
+%         char(model.getMarkerSet().get("head_marker").getAbsolutePathString()));
+    ep.addMocoCoordinateAccelerationGoal('addMocoCoordinateAccelerationGoal',10.0,...
+        div_disp, div_dur, div_mass,2,...
+        {char(model.getCoordinateSet().get(0).getAbsolutePath().toString),...
+        char(model.getCoordinateSet().get(0).getAbsolutePath().toString)}, ...
+        [1.0,1.0]);
 elseif goal_type ==3
     ep.addMocoBOSGoal('base_of_support',10.0,...
         div_disp, div_dur,div_mass, 1,...
