+++ paper/paper.md
@@ -34,13 +34,13 @@ bibliography: paper.bib
 
 # Summary
 
-MocoExtendProblem (`MEP`) is a framework to rapidly develop novel goals for biomechanical optimal control problems using OpenSim Moco [@Dembia2020] and MATLAB (The Mathworks, Inc., Natick, MA, USA). `MEP` features several templates for testing and prototyping novel MocoGoals as well as a build tool to create a MEX function using OpenSim's API for MATLAB in lieu of rebuilding OpenSim from source or building a plugin and generating an .omoco file from C++ to load the problem into MATLAB. Instead, users structure and design custom goals in C++, build them with the provided tool, and call custom goals from within MATLAB scripts.
+MocoExtendProblem (`MEP`) is a framework to rapidly develop novel goals for biomechanical optimal control problems using OpenSim Moco [@Dembia2020] and MATLAB (The MathWorks, Inc., Natick, MA, USA). `MEP` features several templates for testing and prototyping novel MocoGoals as well as a build tool to create a MEX function using OpenSim's API for MATLAB in lieu of rebuilding OpenSim from source or building a plugin and generating an .omoco file from C++ to load the problem into MATLAB. Instead, users structure and design custom goals in C++, build them with the provided tool, and call custom goals from within MATLAB scripts.
 
 This repository features:
 
 - A `build.m` script that compiles goals in the `custom_goals` directory and procedurally constructs the C++/MATLAB class implementations and compiles the MEX interface.
 - Compatibility tested with OpenSim 4.2-4.5.
-  - Support for OpenSim versions 4.2-4.4 require unique considerations to custom goal development and build pipeline since booleans for division by duration, distance and mass were migrated to the abstract MocoGoal. 
+  - Support for OpenSim versions 4.2-4.4 require unique considerations to custom goal development and build pipeline since Booleans for division by duration, distance and mass were migrated to the abstract MocoGoal. 
 - The ability to include `MEP` as a submodule, build, and use valid custom goals.
 - Three example custom goals in the `custom_goals` and `custom_goals_compat` directories.
 
@@ -48,17 +48,17 @@ This repository features:
 
 OpenSim is an open-source software platform for modeling musculoskeletal structures and creating dynamic simulations of movement [@delp2007,@Seth2018]. OpenSim enables researchers and clinicians to investigate how biological and non-biological structures respond to different loads, postures and activities in both static and dynamic situations. OpenSim has been used to study a wide range of biomechanical problems, such as the mechanics of walking and running [e.g. @Falisse2019], the impact of injury or disease on movement [e.g. @Johnson2022], and the effectiveness of rehabilitation exercises [e.g. @Spomer2023].
  
[…]
