### 2025-04-26 #7110 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
